+++ packages/lexical/src/LexicalNodeState.ts
@@ -231,7 +231,7 @@ export interface StateValueConfig<V> {
  */
 export type StateConfigPair<K extends string | symbol, V> = readonly [
   StateConfig<K, V>,
-  ValueOrUpdater<V>,
+  V,
 ];
 
 /**
@@ -277,7 +277,7 @@ export class StateConfig<K extends string | symbol, V> {
    * Convenience method to produce a tuple of a StateConfig and a value
    * of that StateConfig (skipping the parse step).
    */
-  pair(value: ValueOrUpdater<V>): StateConfigPair<K, V> {
+  pair(value: V): StateConfigPair<K, V> {
     return [this, value];
   }
 }
